* Sort other languages links * Option to disable caching
git-svn-id: https://daluang.googlecode.com/svn/trunk@54 ad74ccad-ba46-0410-8d25-6dc0b50a7408

+++ daluang/cache.py
@@ -9,12 +9,25 @@ import time
 class Cache:
 	
 	def __init__(self, cachefile=None, maxsize=500000):
-		if cachefile == None:
-			config = Config()
-			config.init()
-			cachefile = config.read("cache", "/tmp/.daluang-%d.cache" % os.getuid())
+		config = Config()
+		config.init()
 
-		self.cachefile = cachefile
+		# State
+		self.enabled = True
+		enabled = config.read("caching", "True").lower()
+		if not enabled in ["1", "true", "enable", "enabled", "on"]:
+			self.enabled = False
+
+		if not self.enabled:
+			return
+
+		# Cache file
+		if cachefile == None:
+			cachefile = config.read("cache", '/tmp/.daluang-%u.cache')
+
+		self.cachefile = cachefile.replace('%u', str(os.getuid()))
+
+		# Size
 		self.maxsize = maxsize
 
 		self.db = sqlite3.connect(self.cachefile)
@@ -32,6 +45,9 @@ class Cache:
 		return sha.new(wiki).hexdigest()
 
 	def get(self, wiki):
+		if not self.enabled:
+			return None
+
 		hash = self.get_hash(wiki)
 		self.dbc.execute("SELECT hash, timestamp, content FROM cache WHERE hash=?", (hash,))
 		row = self.dbc.fetchone()
@@ -43,6 +59,9 @@ class Cache:
 		return None
 
 	def store(self, wiki, html):
+		if not self.enabled:
+			return
+
 		hash = self.get_hash(wiki)
 		timestamp = int(time.time())
 		cache = html
@@ -57,6 +76,9 @@ class Cache:
 		self.db.commit()
 
 	def remove(self, wiki):
+		if not self.enabled:
+			return
+
 		hash = self.get_hash(wiki)
 		self.dbc.execute("DELETE FROM cache WHERE hash=?", (hash,))
